+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 1.1.9 - 2013-05-02
 seaweedchan
 - Fix boards with previously deleted archives not switching to new archives